BREAKING CHANGE: IST-2638: Move clip window to job from sources
Clip window is no longer a valid input for sources. Instead it can now (optionally) be provided to a singlecam/multicam job, or for volume creation. If not provided the entire video is processed.
