+++ docs/api-reference/schemas/job.md
@@ -2,4 +2,10 @@
     options:
         members: ["id", "created", "metadata", "client", "outputs", "state", "take"]
 
+::: move_ugc.schemas.job.ClipWindow
+    options:
+        members:
+            - start_time
+            - end_time
+
 ::: move_ugc.schemas.job.JobOptions
